Fix and add style for most of the <form> elements
Fixed the missing attributes from the <form> children elements

Fixed almost all the <form> elements CSS thanks to a better
undestanding of flexbox allignment/justify properties mostly due to
https://css-tricks.com/snippets/css/a-guide-to-flexbox/

Changed the font icon style from Round to Sharp since it seems to match
more closely the original page, the only one that should be round is
the "apps" one but appears squared even in their site (i may have a
temporary workaround for that using repeated different icons)

Added a "Color Palette" linking specific colors to variables for
consistency across the page / site
The ones named as --clr-darker-XX set the color to black with 0.XX
as the alpha value, creating a darker shade that should work even if
the background color changes

Removed most of the CSS for the header and footer, it was either poorly
implemented or just plain wrong. Planning to work on those on separate
individual braches to practice splitting and merging the code in a
project.

Added some commented lines starting with "???" for question to get
clarification on after the end of the project (will probably add
reference to those in the README.md for ease of access later on)