Fix "carbon neutral" banner position and wrapping
Centering:
Sets the min-width of the other siblings in the footer to a width in
rem units thats >= of the largest of the 3, this leaves a now centered
space in the middle where foot__eco itself can be centered with
margin: 0 auto;

Wrapping:
The media query max-width value is set to the value of min-width chosen
for the Centering multiplied for the number of elements, this ensures
that the flex container will never wrap before the media query is
active. Then the foot__eco is given a flex order lower than its sibling
and a width of 100vh to force a wrap that puts it on top of them.