Add a workaround for issue #5
styles/style.css Lines 113-114

Line 113 leaves the main section close to  where it should be for the
most popular desktop screen sizes.

Line 114 is the value from the original google.com homepage, when is
used the behavior is closer to the original but the main section sits
too high on the page.